refactor: share one Turso DB across wack-hacker; prefix shopping tables
Rename env vars to the default TURSO_DATABASE_URL / TURSO_AUTH_TOKEN since this
Turso instance is for all wack-hacker features, not just shopping. Move the
drizzle client up to src/lib/db.ts so non-shopping features can share it.
Prefix tables as shopping_carts / shopping_cart_items (and the unique index)
so future feature tables can coexist in the same database without collisions.
Migrations now live at drizzle/ (shared history).
