+++ .env.example
@@ -39,9 +39,9 @@ GROQ_API_KEY=
 SHIP_DATABASE_TURSO_DATABASE_URL=
 SHIP_DATABASE_TURSO_AUTH_TOKEN=
 
-# Shopping cart database (Turso / libsql)
-SHOPPING_DATABASE_TURSO_DATABASE_URL=
-SHOPPING_DATABASE_TURSO_AUTH_TOKEN=
+# Shared wack-hacker database (Turso / libsql)
+TURSO_DATABASE_URL=
+TURSO_AUTH_TOKEN=
 
 # Cloudflare R2 (image storage)
 R2_ACCOUNT_ID=
