+++ drizzle.config.ts
@@ -1,16 +1,16 @@
 import { defineConfig } from "drizzle-kit";
 
-const databaseUrl = process.env.SHOPPING_DATABASE_TURSO_DATABASE_URL;
+const databaseUrl = process.env.TURSO_DATABASE_URL;
 if (!databaseUrl) {
-  throw new Error("Missing required environment variable: SHOPPING_DATABASE_TURSO_DATABASE_URL");
+  throw new Error("Missing required environment variable: TURSO_DATABASE_URL");
 }
 
 export default defineConfig({
   dialect: "turso",
   schema: ["./src/lib/shopping/schemas/*.ts", "./src/lib/shopping/relations.ts"],
-  out: "./drizzle/shopping",
+  out: "./drizzle",
   dbCredentials: {
     url: databaseUrl,
-    authToken: process.env.SHOPPING_DATABASE_TURSO_AUTH_TOKEN,
+    authToken: process.env.TURSO_AUTH_TOKEN,
   },
 });
